Fix Bugs
A few bugs were fixed with displaying items and error handling. In weather.js, an error thrown when reaching out to the API will now hopefully propagate up to the calling function so that the caller can send back a proper response back to the client.

A few style changes were also made so that offices correctly wrap in WeatherWidget box.