Attempt to fix bugs relating to testing
Once tests were added to the backend, starting the front end was causing problems. The front end could detect that jest had been installed, and was reporting issues that it couldn't find a version of jest it was compatible with. Jest was installed on the front end through npm. A setting in a .env file is what ended up fixing the issue in the end.